balancergroup: convert tests to blackbox (package balancergroup_test)
Convert balancergroup_test.go in-place from whitebox (package
balancergroup) to blackbox (package balancergroup_test) testing.
This minimizes the PR diff by keeping the same filename and only
changing the lines that need to differ for external package access.

Changes:
- package balancergroup → package balancergroup_test
- Add imports: "strings", "google.golang.org/grpc/internal/balancergroup"
- Qualify types: New → balancergroup.New, Options → balancergroup.Options,
  BalancerGroup → balancergroup.BalancerGroup, ParseConfig → balancergroup.ParseConfig
- defaultTestShortTimeout: 10ms → 100ms (avoid flakes)
- Move TestBalancerGroup_RemoveImmediately from ext_test into main test file
- Delete balancergroup_ext_test.go (no longer needed)
